requirements.txt
pika
nose
ptree
ConcurrentLogHandler
lxml
BeautifulSoup
dateutils
httpretty
requests
supervisor
